+++ pageflash.php
@@ -35,7 +35,10 @@ define( 'PAGEFLASH_PATH', plugin_dir_path( PAGEFLASH_FILE ) );
 define( 'PAGEFLASH_URL', plugins_url( '/', PAGEFLASH_FILE ) );
 define( 'PAGEFLASH_ASSETS_PATH', PAGEFLASH_PATH . 'assets/' );
 define( 'PAGEFLASH_ASSETS_URL', PAGEFLASH_URL . 'assets/' );
+define( 'PAGEFLASH_BUILD_PATH', PAGEFLASH_PATH . 'build/' );
+define( 'PAGEFLASH_BUILD_URL', PAGEFLASH_URL . 'build/' );
 define( 'PAGEFLASH_ENV', WP_DEBUG ? 'development' : 'production' );
+define( 'PAGEFLASH_ICON', PAGEFLASH_URL . 'assets/logo/icon.svg' );
 
 add_action( 'plugins_loaded', 'pageflash_load_plugin_textdomain' );