+++ src/admin/actions/actionTypes.js
@@ -1,0 +1,2 @@
+export const ACTIVE_MENU = 'ACTIVE_MENU';
+export const SET_LANDMARKS = 'SET_LANDMARKS';