+++ src/admin/components/layout/WithHeaderLayout.jsx
@@ -1,0 +1,13 @@
+import { Outlet } from 'react-router';
+import { Header } from '../header';
+
+const WithHeaderLayout = () => {
+	return (
+		<Header>
+			<main className="pageflash-main-content">
+				<Outlet />
+			</main>
+		</Header>
+	);
+};
+export default WithHeaderLayout;