+++ src/admin/context/index.js
@@ -1,0 +1,9 @@
+import { createContext } from 'react';
+
+/**
+ * @constant {React.Context} AdminContext - A context object created using React's createContext API.
+ * Initialized with a default value of null.
+ */
+const AdminContext = createContext( null );
+
+export { AdminContext };