+++ src/admin/provider/AdminProvider.jsx
@@ -1,0 +1,22 @@
+/* eslint-disable react/prop-types */
+import { useReducer } from 'react';
+import { AdminContext } from '../context';
+import { initialState, rootReducer } from '../reducers';
+/**
+ * AppProvider component that sets up the application context and routing.
+ *
+ * @param {Object} props - The properties object.
+ * @param {React.ReactNode} props.children - The child components to be wrapped by the provider.
+ * @returns {JSX.Element} The provider component with routing and context.
+ */
+const AdminProvider = ( { children } ) => {
+	const [ state, dispatch ] = useReducer( rootReducer, initialState );
+
+	return (
+		<AdminContext.Provider value={ { ...state, dispatch } }>
+			{ children }
+		</AdminContext.Provider>
+	);
+};
+
+export default AdminProvider;