+++ src/admin/utils/capitalizeFirstLetter.js
@@ -1,0 +1,7 @@
+export const capitalizeFirstLetter = ( string ) => {
+	if ( typeof string !== 'string' || string.length === 0 ) {
+		return string;
+	}
+
+	return string.charAt( 0 ).toUpperCase() + string.slice( 1 );
+};