+++ src/admin/utils/index.js
@@ -1,0 +1,6 @@
+import { capitalizeFirstLetter } from './capitalizeFirstLetter';
+import { createRandomId } from './createRandomId';
+import { filterLandmarksByMenu } from './filterLandmarksByMenu';
+
+export { capitalizeFirstLetter, createRandomId, filterLandmarksByMenu };
+